--- GET
@@ -1,16 +1,0 @@
-GET / HTTP/1.1
-Host: localhost:9997
-Connection: keep-alive
-Upgrade-Insecure-Requests: 1
-User-Agent: Mozilla/5.0 (X11; Linux x86_64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/86.0.4240.111 Safari/537.36
-Accept: text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.9
-Sec-Fetch-Site: none
-Sec-Fetch-Mode: navigate
-Sec-Fetch-User: ?1
-Sec-Fetch-Dest: document
-Accept-Encoding: gzip, deflate, br
-Accept-Language: fr-FR,fr;q=0.9,en-US;q=0.8,en;q=0.7
-If-Modified-Since: Fri, 30 Oct 2020 14:23:03 GMT
-
-
-Host: you should not see this
